backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.23
pydantic>=2.6
pydantic-settings>=2.2
sqlalchemy>=2.0
asyncpg>=0.29
alembic>=1.12
cryptography>=42.0
argon2-cffi>=23.1
PyJWT>=2.8
python-multipart>=0.0.9
boto3>=1.28
redis>=5.0
celery>=5.3
prometheus-client>=0.17
reportlab>=4.0
typer[all]>=0.9
httpx>=0.25
jinja2>=3.1
orjson>=3.9

# Testing
pytest>=7.4
pytest-asyncio>=0.21
requests>=2.31
aiosqlite>=0.19
anyio>=4.3
